refactor query log
